fix, update quotation entity

+++ src/main/java/online/noithat/be/Entity/QuotationDetail.java
@@ -21,7 +21,7 @@ public class QuotationDetail {
     @JsonIgnore
     Quotation quotation;
 
-    Float quantity;
+    long quantity;
     Float price;
 
 }
